+++ src/main/java/org/tonkushin/hw05/dao/book/BookMapper.java
@@ -11,14 +11,14 @@ import java.sql.SQLException;
 public class BookMapper implements RowMapper<Book> {
     @Override
     public Book mapRow(ResultSet resultSet, int i) throws SQLException {
-        int id = resultSet.getInt("books.id");
+        long id = resultSet.getLong("books.id");
         String name = resultSet.getString("books.name");
 
-        int authorId = resultSet.getInt("author_id");
+        long authorId = resultSet.getLong("author_id");
         String authorName = resultSet.getString("authors.name");
         Author author = new Author(id, authorName);
 
-        int genreId = resultSet.getInt("genre_id");
+        long genreId = resultSet.getLong("genre_id");
         String genreName = resultSet.getString("genres.name");
         Genre genre = new Genre(id, genreName);
